requirements.txt
google-cloud-bigtable==2.10.1
google-cloud-core==2.3.1
